move some assistant code to /cloud and to /xplat
