+++ CopilotKit/packages/cloud/src/index.tsx
@@ -1,1 +1,2 @@
+export * from "./streams";
 export * from "./types";
