+++ CopilotKit/packages/cloud/src/streams/index.ts
@@ -1,0 +1,1 @@
+export { experimental_AssistantResponse } from "./assistant-response";
