--- CopilotKit/packages/cloud/src/utils/utils.test.ts
@@ -1,7 +1,0 @@
-import * as utils from "./utils";
-
-describe("emptyTest", () => {
-  it("should be truthy", () => {
-    expect(true).toBeTruthy();
-  });
-});
