--- CopilotKit/packages/react-core/src/openai-assistants/types/index.ts
@@ -1,1 +1,0 @@
-export type { Message } from "./shared";
