+++ CopilotKit/packages/react-core/src/openai-assistants/utils/index.ts
@@ -1,0 +1,1 @@
+export { processMessageStream } from "./process-message-stream";
