+++ CopilotKit/packages/xplat/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./hooks";
+export * from "./types";
 export * from "./utils";
