--- CopilotKit/packages/xplat/src/index.tsx
@@ -1,2 +1,0 @@
-// export * from "./types";
-export {};
