--- CopilotKit/packages/xplat/src/types/dummy.ts
@@ -1,1 +1,0 @@
-export interface SampleInterface {}
