+++ CopilotKit/packages/xplat/src/types/index.ts
@@ -1,1 +1,1 @@
-export type { SampleInterface } from "./dummy";
+export * from "./openai-assistant";
